HtmlBridge Phase 3 (P3.15): co-locate the CSSOM stylesheet/rule object model
The CSSOM CSSRuleList/CSSRule object model — the sibling of P3.14's
CSSStyleDeclaration — is now the fifteenth co-located feature module,
StyleSheetBinding (namespace Broiler.HtmlBridge.Dom.Features, split
.cs/.Rules.cs/.Callbacks.cs to stay under the 750-line/file guideline). It
owns the rule-list + keyframe builders, the per-rule CSSRule builder for
every rule kind (style/@media/@supports/@layer/@keyframes/@font-face/@page/
@property/@counter-style/@import/@namespace, via the neutral
Broiler.CSS.Cssom.CssomRuleMetadata projection), and the ~20 JsStyleSheets*Core
callbacks (length/item/cssRules/insertRule/deleteRule + the per-kind cssText
serializers).

Like ClassListBinding (P3.6) and StyleDeclarationBinding (P3.14) it is an
internal static class with no host contract — pure CSSOM-IDL logic over the
shared Broiler.CSS rule model, CssParser/CssSerializer, and
StyleDeclarationBinding.BuildRuleDeclaration; the one bridge helper it needs is
the neutral static DomBridge.ParseStyle. The CSSStyleSheet object itself
(per-element identity cache, live cssRules collection, insertRule/deleteRule
mutation bookkeeping) stays bridge-owned in DomBridge.BuildStyleSheetObject
(runtime-state coupled) and calls into the module.

Moves all rule-object code out of the 918-line StyleSheets.cs (now 141) and
the deleted JsFunctionCallbacks/StyleSheets.cs, bringing StyleSheets.cs under
the Phase 3 750-line exit-criterion limit. Behaviour-preserving; no public-API
change (module + callbacks internal).

Tests: Broiler.Cli.Tests/StyleSheetBindingModuleTests.cs (co-location/static
guard + selectorText/rule.style/cssText, insertRule/deleteRule live-collection
mutation, @media/@keyframes rule-kind characterizations through the JS engine).
Regression: the SelectorsAndCssom / StyleDeclarationBinding / Acid3CssCompliance
/ CssStyleDeclarationValidation / WptCssVariables and architecture-guard suites
reproduce an identical failure set with the change stashed (the standing :lang,
Acid3 border-shorthand, and css-variable-shorthand environmental fails) — zero
regressions.
